+++ src/main/java/tenno_mod/TennoMod.java
@@ -98,6 +98,7 @@ public class TennoMod implements EditCharactersSubscriber,
     cardsToAdd.add(new Meditation_TENNO());
     cardsToAdd.add(new BlindingSlice_TENNO());
     cardsToAdd.add(new QuickSlash_TENNO());
+    cardsToAdd.add(new QuickGuard_TENNO());
     cardsToAdd.add(new AgileGuard_TENNO());
     cardsToAdd.add(new DefensiveThrow_TENNO());
     cardsToAdd.add(new FallingKick_TENNO());